Squashed Networkpolicylogging commits including following:
1. Add NetworkPolicyLogging API definition

2. Add scripts to autogenrate API clientset and deep-copy methods

Auto generated clientset and deepcopy methods for network policy logging

3. Fix the scope of the policylogging PRD to be non-namespaced.

Also fix the license

Bug: b/153083245

4. Create a NetworkLogging CRD to configure loggings.

A singleton CRD is used to configure logging for network policy,
node network policy and flow logs.

5. Add auto-generated files for NetworkLogging CRD.

6. Add the flow dispatcher.

Flow dispatcher observes the hubble flow event and pass it
to any gke listener that registerd.

7. Remove the old networkpolicylogging CRD definition.

Also removed the changes we made to the hack directory.

8. Support network policy action logging.

This patch implements the network policy action logging.
Design doc: go/gke-nwp-logging-design
User guide: go/gke-nwp-logging-user-guide
Code structure diagram: go/gke-logging-block-diagram

9. Insert GKE plugin to hubble work flow

This serves as a workaround to insert GKE plugin before Atlantis
build system is ready.

10. Add event recorder vendor

This vendor change is needed for the PR
https://gke-internal-review.googlesource.com/c/third_party/cilium/+/150005

Separate the commit is just for easier reviewing.
The change is done via "go mod vendor", but then the change cannot be committed
to GOB because the OWNER file fails the GOB format check (b/160597845).
So I renamed the file to GIT_OWNER to bypass the issue.

11. Report k8s event when updating NetworkLogging CR.

Fix: b/158724727

12. Adjust the max log rate for policy logging

We talked to logmon team and decided to support 500 logs per second
for GKE on GCP.

13. Skip flow events generated by node network policies.

Policy logging is not supported yet, thus ignore the events
generated from it for now.

Bug: b/166676540

(cherry picked from commit e9cecae0ce17ffd26cce530c20f5b55b5783a3ed)

14. Remove the old counter implementation.

Will add the new metrics based on the proposal: go/gke-policy-logging-metrics

Bug: b/166812557

15. Move the test constant to test file for policy_logger test.
16. Modify the policy logger start logic to fit in the ready metric.

This change is part of implementing the readiness SLI as described
in go/gke-policy-logging-metrics.

The policy logging feature is considered to be ready if both the
controller cache is synced and the policy logger pipeline are
setup correctly. The change sets up the policy logger pipeline at
the node start time, instead of being controlled by the NetworkLogging
CR. Policy events will be dropped early if logging is disabled.
This is aligned with our long-term vision that counter should be
supported regardless if logging is enabled or not.

Bug: b/166812557

17. Add policy logging metrics definition

The metrics is defined based on the proposal: go/gke-policy-logging-metrics

Bug: b/166812557

18. Set the values for the prometheus metrics.

The example output can be found in go/gke-policy-logging-metrics

Bug: b/166812557

19. Add resource version to policy logging enable/update event

Bug: b/163421262

20. Add the internal log rate configuration

This was missed in the initial commit. It is not exposed to customer
and it allows us to increase the rate limit easily in performance tests.

21. Fix clientset scheme for nodenetworkpolicy events

BUG: b/155193289

(cherry picked from commit deebb6accc51b997993d7733881cd98e5ee3b268)

22. Fix policy correlation workflow some edge cases

Policy correlation does not as expected for the following two cases:

1. A policy that allows all traffic on a specific port-protocol combination.

In this case, one or multiple identities may be allowed as label selectors
are not specified. So, the endpoint's policy map (remote identity -> policy)
will have the identity field in the key set to unknown-identity (all identities).
But the hubble flow always come with a known single destination identity
and we use that to retrieve the policy. So, we fail to retrieve the policy.

For example, A policy that allows all traffic on TCP port 80 will run into this.

2. A policy that allows all traffic of a specific protocol (for the
same reason as above)

BUG: b/182422598

23. internal workload kind and name changes

24. policy_event metric exporter

This change adds a policy_event metric exporter that collects policy_event_count metric with label fields:
namespace, pod_name, verdict, workload_name, workload_kind and direction. This would eventually be exported by gke-metrics-agent as
an external gke metric that would be visible to customers by default.

Commits
- adding skeleton code for metrics exporter
- added flow listener, updated controller, added a new metric and renamed packages
- implemented extraction of metric fields from flow and did minor clean up
- did some refactoring, improved processFlowAndRecordMetric method and added unit test
- cleaned up processFlowAndRecordMetric function and updated unit test
- changed flowListener name to decouple from network policy logger
- made verdict and direction lowercase
- updated unit test
- made changes to incorporate review comments
- removed api interface and changed namespace field in metric initialization
- updated code to consider static podsi
- fixed a bug after load testing where flow was getting dropped due to a full queue
- added a separate stop channel for metric exporter in plugin, used
cmp.diff in unit tests and did minor refactoring
- ran golint and made recommended changes

go/gke-nwp-action-metrics-wo-logging-requirements
go/pol-event-count-impl-design
go/clearance_for_new_public_metric
go/network-policy-event-metric-load-test-results
b/194398494

25. Add back the pod_namespace to log entries

Bug: b/197769112

26. updated namespace field to avoid conflict with existing namespace field

b/198219587

27. renamed pod & namespace field labels for policy event count metric

28. feat: add network logging support for cilium
network policies.

example log:
https://paste.googleplex.com/4806276511956992

- add delegate support for cilium np CRs

BUG: b/246835150
BUG: b/245849745

28. Remove duplicate policies in policy action logs

The issue occurs when two policies (one that selects
a specific pod labels and another that selects all
pods in the namespace) select a specific workload.

BUG: b/255548024

29. fix: support correlation for cluster-scoped APIs.

This change allows correlation for cluster-scoped resources.

The function does not validate whether the resource *is* namespaced.

Bug: b/287232684

30. fix: reduce log spam.

This condition is triggered for transparent policies that are are
not derived from a policy CR.

31. fix: use different path for every logger test.

Change-Id: I80b27e82eca1583efef7eb6026d7848263b4b17c
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/778681
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Akhil Velagapudi <avelagap@google.com>