Add event recorder vendor
This vendor change is needed for the PR
https://gke-internal-review.googlesource.com/c/third_party/cilium/+/150005

Separate the commit is just for easier reviewing.
The change is done via "go mod vendor", but then the change cannot be committed
to GOB because the OWNER file fails the GOB format check (b/160597845).
So I renamed the file to GIT_OWNER to bypass the issue.

Change-Id: Ie17c1ed74f97f8feff48f336afa1ff70b584999a