Add node identity to the allowed list if allowed CIDR contains local node IP
This provides a workaround for node cidr issue
https://github.com/cilium/cilium/issues/12277
when the CIDR of a network policy includes the whole node cidr range.
Local host and remote nodes are not distinguished to make the workaround simple.

Note that if the CIDR of the network policy isn't larger than node CIDR, this patch will
cause inconsistent behavior across nodes. Also although the patch is intended for
Kubernetes network policy, it would change the behavior of Cilium network policies
as well as they share the same code path.

This solution is described in detail in go/cilium-network-policy-cidr-issue

Bug: b/160334450

Change-Id: Ie9ceccf12cd92fd1d94b756f539a8b5cd6ca7933
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/785732
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Satish Matti <smatti@google.com>
Reviewed-by: Akhil Velagapudi <avelagap@google.com>