Add redirect service.
This a squashed commit that contains the following changes:

Implements RedirectService CRD for nodelocaldns.

pkg/gke/apis/* and pkg/gke/client/* are generated code, no need to review those :)

daemon/cmd/daemon.go: hook point to start CRD controller

pkg/gke/redirectservice/controller/controller.go: main controller
functions

pkg/gke/redirectpolicy/gke_helpers.go: main helper function from Cilium
side

Design doc: go/dpv2-nodelocaldns
BUG: 156649951
[I49e8281ba983bbc734b1c7cb85cbcc51067b5b83]

Adds an option to control whether to turn on/off redirect service.

This is not needed for on-prem now so disable it to avoid red herring
in the cilium-agent log.

[I6de247153faea0b6dbeb35abad7f750562041135]

Reacts to pod deletion locally for nodelocaldns

Current implementation of LRP relies on k8s control plane to react to
backend pod events. In large clusters, this causes a delay when
restoring DNS service to kube-dns upon the event of node-local-dns pods
(e.g., a rolling upgrade of node-local-dns) and this delay will lead to
DNS queires lost because the agent will still direct DNS packets to the
local backends which no longer exists and will only stop doing so when
it got the local backend deletion event from the api-server.

This PR invokes the service restoration logic in the local pod deletion
path. When kubelet issues cmdDel to Cilium-CNI, we still proactively
hint redirect service manager and let it decide whether we need to
restore original svc or not.

Manually tested in a small cluster.

BUG: 189861453
[I559c3b062e368e731185d30221531f68f08ddaf8]

Check for endpoint.pod exsitence before removing

When restrarting agent, if the node-local-dns pod does not exist for
some reason, we do not want to "remove" it from the LRP backend maps
because:
- it won't exist there
- endpoint.pod would be nil

BUG: 199771679
[I6aeaa348546f7b81cf74bcbad46fa97dd687a93a]

Adds "cloud-dns" provider to redirectservice manager

BUG: 236415610
[Id44a7e5a3109cbb70e0ea07328aa084b29c25ed7]

Allows for temp NLD ds to skip conntrack too

This is an optimization that was missed for the temp nodelocaldns instance when doing DPv2 migration. Missing this optimization won't break DNS and nodelocaldns functionality but keeping the local connection between pods and node-cache in kernel conntrack (not ideal but should be okay in most cases).

[Ic669f6325596419bc6bff23a399aa061794b2441]

Change-Id: I81445b6ef790039d41d1f29227c2343420294270
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/782722
Reviewed-by: Akhil Velagapudi <avelagap@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>