Fix policy correlation workflow some edge cases
Policy correlation does not as expected for the following two cases:

1. A policy that allows all traffic on a specific port-protocol combination.

In this case, one or multiple identities may be allowed as label selectors
are not specified. So, the endpoint's policy map (remote identity -> policy)
will have the identity field in the key set to unknown-identity (all identities).
But the hubble flow always come with a known single destination identity
and we use that to retrieve the policy. So, we fail to retrieve the policy.

For example, A policy that allows all traffic on TCP port 80 will run into this.

2. A policy that allows all traffic of a specific protocol (for the
same reason as above)

BUG=b/182422598

Change-Id: I780c2c79caf148fcd565fcfef951501a8e82c40b