Restore identity for tunnel mode when packet is from Windows
This change fixes the issue where ingress policy on Linux pod is broken
when remote client is Windows pod, in tunnel mode. The issue happens
because cilium users tunnel id to carry pod identity, which will be used
to enforce network policy. If remote is Linux node, cilium agent will
encode the correct identity in tunnel id. If remote is Windows, the
information is missing. In this case, we need to recover the identity by
looking up ipcache using source IP address, the same as direct routing
mode where tunnel id is not available.

Fix: 217232644
Change-Id: I73631bb680e69dfaae9d809fada8cad535522a90