Add ability to limit clustermesh to certain namespaces.
The clustermesh-apiserver takes in configuration of a set of labels.
If configured, then clustermesh only distributes services, endpoints and
identities in namespaces that have these labels. In addition, we don't
distribute nodes when namespace label restrictions are configured. If not
configured, then we keep the default behavior of distributing everything.

BUG: 213524952

Change-Id: Ib098ac251d64ccdfe932f5492bbc03c9ebfa8ac2