Fix multi-network NodePort for local backends
For MACVLAN interfaces, when a frame enters the lower device, if the
destination MAC is not one of the sub-interfaces, the frame is processed
by the host. This breaks isolation, and leads to a strange datapath
where the kernel routes the packet back to the lower device where it's
captured and redirected back to the ingress-side (by
`multinic_redirect_ipv4`) for another attempt at reaching the pod.

To avoid this, set the correct destination MAC after xlating the
destination addr for multi-NIC endpoints.

Also, redirecting nodeport egress packets to host interface by the
`from-container` program doesn't work for MACVLAN because the program is
running in the pod namespace. As a workaround, don't redirect and rely
on the MACVLAN driver to properly egress the packet.

Fix: b/233754304
Change-Id: Id6165b129ad5fa311be8c0774b587f9347c50663
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/542981
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Varun Marupadi <varunmar@google.com>
Reviewed-by: Maciej Skrocki <maciejskrocki@google.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/785975
Reviewed-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Akhil Velagapudi <avelagap@google.com>