Set tunnel_id to 0 if id is larger than 24 bits.
IDs larger than 24 bits are CIDR identities which are local to the node
so should not be send in a tunnel ID on the wire. Attempting to send
them means they get truncated and the receiver interprets the ID
differently than intended. This is an issue for the return path of the
egress NAT traffic flow only.

BUG: 239603692
Change-Id: I6a08b3bde8b5bb5778062e795726760abe8fdfa0
(cherry picked from commit 35cb5fe60bacd2e3caf29172fbab5671897f12ad)