Allows for temp NLD ds to skip conntrack too
This is an optimization that was missed for the temp nodelocaldns instance when doing DPv2 migration. Missing this optimization won't break DNS and nodelocaldns functionality but keeping the local connection between pods and node-cache in kernel conntrack (not ideal but should be okay in most cases).

Change-Id: Ic669f6325596419bc6bff23a399aa061794b2441
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/537620
Reviewed-by: Dmitry Pasiukevich <dpasiukevich@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Varun Marupadi <varunmar@google.com>