[REBASE/REWRITE MATCHED WITH MASTER]
The commits between upstream and this marker have all been included
in the master branch up to the corresponding commit called:
[REBASE/REWRITE MATCHED WITH v1.12]
We will use this commit to keep track of new changes that may need
to be cherrypicked in either direction.

Commits MUST NOT be rebased or moved across this boundary for the
life of this branch.

Change-Id: I978ccf53a2fc060e4c9619c814dba6142065f713
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/549444
Reviewed-by: Weilong Cui <cuiwl@google.com>
Reviewed-by: Maciej Skrocki <maciejskrocki@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>