IPAM changes to handle multinetwork pods.
Creates a hostScopedAllocator per network by parsing the network
annotations on the node. The created allocator is saved in a
network-allocator map from which the it is referenced. The allocator
then allocates the next IP and releases the allotted IP upon endpoint
deletion. CL also includes changes to the reconciler.

go/multinet-ipam

Change-Id: Ifd20982464b47f32fbc7a8a26b34c985c76c5926
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/494465
Reviewed-by: Yuan Liu <liuyuan@google.com>
Reviewed-by: Yifeng Shen <yfshen@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/558262