Skip service map generation for non "Running" pods
This is to address a bug where deleting a completed pod would
delete the service map entries for a running pod on the same node, due
to the hostPort mapping key being the same for the old and the new pod.
Intuitively, a completed pod to shouldn't affect the service map entries
on a node, so it makes sense to skip generation, in order to prevent any
interference with entries.

BUG: b/260110416
Change-Id: I3a3f76a2aec794cc5b5ae216ca418b72d8a79d33
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/581322
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Maciej Skrocki <maciejskrocki@google.com>