Don't masquerade on overlay interface in flat mode.
This addresses an issue with egress NAT where the node was NATing to the
egress NAT IP on the overlay interface of the sender node. In flat IP
node we never have to NAT on the overlay interface which is only used
for egress NAT tunelling.

BUG: 261679736
Change-Id: Ie868df227683428937995900be9b2db6e2c929a2
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/582910
Reviewed-by: Sarvesh Rangnekar <sarveshr@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/584710