Backport https://github.com/cilium/cilium/pull/22865
hubble: Add a metric for lost events

Currently numbers of lost events are visible in logs and in the observer, but
they tend to be noisy and difficult to consume in such form. A Prometheus
counter is much more suitable for this kind of information, thus this patch
adds hubble_lost_events_total metric.

Fixes #15112

hubble: Log number of lost messages on Debug level

The number of lost messages is now exposed as a metric, which is more practical
than a log, so the log level can be reduced. Also rephrased the log message, to
indicate only the fact that a LostEvent message was received - quite often the
queue will be full immediately after that, so saying that Hubble is "back to
nromal" can be misleading.

docs: Document hubble_lost_events_total metric

Signed-off-by: Anna Kapuscinska <anna@isovalent.com>
Change-Id: I7a4d78c53dacb444a236f95a6301c221216f2300
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/808520
Reviewed-by: Alan Kutniewski <kutniewski@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>