[backport from master] bpf: nodeport: rename *_ENCAP_NODEPORT_NAT tail-calls
We have
1. NODEPORT_NAT_EGRESS for SNAT of requests that get forwarded by the LB,
2. NODEPORT_NAT_INGRESS for revSNAT / revDNAT of replies that enter the LB.

Pick NODEPORT_NAT_FWD for the path where local traffic leaves the node
via to-netdev or to-overlay. This is used for BPF Masquerading, and in the
future also for handling replies by a local backend.

Also adjust the naming of some related functions.

Change-Id: Ie99c42ed5621b13ac65de577b02e03d70ad2dfb4
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/613344
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Maciej Skrocki <maciejskrocki@google.com>