Set network-status annotation to empty list explicitly
Since when we started to support both pod-network and default network,
network-status annotation is first set on the node without any value
during the pod-network reconcile and this is causing a parsing error
when default network is reconciled.

Explicitly setting the value to an empty list until we stop supporting
pod-network.

Change-Id: I509c9bbe3f317322d688b863f76999d7dab4907f
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/632830
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Maciej Skrocki <maciejskrocki@google.com>