[backport] EndpointSlice semantics
Backport of commit #1 from upstream PR
https://github.com/cilium/cilium/pull/24174.

Original commit message:

Fix the usage of the EndpointSlice Conditions semantics:
Ready, Serving and Terminating.

The conditions of the Endpoints and EndpointSlices are complicated
because to keep backwards compatibility in the API a new Serving
state has been added.

This Serving state must be used to include and endpoint that is
Terminating, not Serving pods should not be used as backends.

Signed-off-by: Antonio Ojea <aojea@google.com>

Bug: b/275447567
Change-Id: I6bd4b84d2d4c0f4fbbca272c494d625493afef93
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/657078
Reviewed-by: Yash Shetty <yashshetty@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Varun Marupadi <varunmar@google.com>