Reduce log spam for auto-direct-node-routes
Error for installing node routes, when nodes spans across different L2
domains is a benign error and should be logged with a debug level.
Currently it is logged as a warning and spams the logs, making it hard
to spot actual errors.

BUG=267522878

Change-Id: If84707a28dd01a1cc723c6cbb56d7f752968ac88
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/680041
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Jonathan Hart <jonohart@google.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/782954
Reviewed-by: Akhil Velagapudi <avelagap@google.com>