Do not let Cilium manage devices with prefixes provided in --device-prefixes-to-exclude.
With the flag --enable-runtime-device-detection enabled, Cilium tries to
dynamically pick up new devices and load eBPF programs onto them. This
fails in a few cases: 1) for qbone0, which is a privileged device being
used in the tunnel to the Google prod network, and 2) for vti-anthos
devices which are virtual tunnel interfaces used to establish VPN
connections to GCP. We solve this by adding a parameter which excludes
devices with provided prefixes, and passing in the prefixes for the
devices that we want to prevent Cilium from managing.

Bug: b/279040119
Tested: Created custom image and used in GDCE cluster, created and deleted VPN connection several times and did not observe sessions crashing. Also did not observe errors in logs suggesting that Cilium was trying to manage qbone0 or vti-anthos devices.

Change-Id: I3dfb38b6cdb94171f5664b259e092d00a5e8f50c
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/709507
Reviewed-by: Yuan Liu <liuyuan@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/795058
Reviewed-by: Akhil Velagapudi <avelagap@google.com>
Reviewed-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Yash Shetty <yashshetty@google.com>