envoy proxy: Only reuse DNS proxy port when it's free
When cilium-agent starts, it will allocate a free port for proxy to
use, if users don't speicify in config. It also tries to recover
previous allocation from iptables rules, but the recover doesn't check
if the port is already open by other processes on the host. This change
will check the recovered port is free before assign it to DNS proxy.

Fix #22465
Signed-off-by: Yongkun Gui <ygui@google.com>

BUG: 259295873
Change-Id: Ic0e116cdc30c0dd7901ea1c13b2b00a67ea117fa
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/716448
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Akhil Velagapudi <avelagap@google.com>