Increase the max wait time for EndpointCreate API requests
In the event that a node has a burst of PodSanboxCreate requests, API
requests will pile up. However, we should wait longer for the queue to
clear before short-circuting and returing failure.

This is because the kubelet has a relatively relaxed timeout for
PodSandbox creation -- 4 minutes. Furthermore, if we return a failure
here, it is propagated all the way back through containerd to kubelet,
which will tear down the entire PodSandbox and try again, which can be
expensive.

So, increase the maximum queue duration time to 1 minute. That should
hopefully give enough time for the queue to clear.

Upstream-PR: https://github.com/cilium/cilium/pull/25805
Signed-off-by: Casey Callendrello <cdc@isovalent.com>
Change-Id: I19a55908e5d81e5970b0215ca6114780d7d40c56
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/856899
Reviewed-by: Yash Shetty <yashshetty@google.com>
Tested-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>