Add Flags to enable CCNP
There are two flags to toggle CNP and CCNP independently.

BUG=300518289

This was previously published as gkecl/833051, but we missed some
dependent commits in that master branch. So we move to this staged
branch, and it will be in the master
branch once the rebasing is completed.

Change-Id: Ia576e967c62e49ec63b5aac52f0064ea55d69854
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/836308
Reviewed-by: Blaz Zupan <blaz@google.com>
Reviewed-by: Satish Matti <smatti@google.com>
Reviewed-by: Akhil Velagapudi <avelagap@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Tested-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
(cherry picked from commit babd3b5bf4524ef40711ed2095b4144061111eb7)
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/839040