Initialize bootstrap metrics just before populating the data.
Can be dropped in v1.14+

This avoids a race condition where the pkg/metrics/metrics.go init can run later and reset the metrics.

This is an internal-only patch. The change has been fixed in upstream main branch, but that includes several more changes and hence is not easy to backport.


Bug: b/297234936
Change-Id: I30cb69f88b9759742292a8445af18cc4382e52c7
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/864572
Reviewed-by: Akhil Velagapudi <avelagap@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Tested-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>