Skip checking precondition for host endpoint label updates
This fixes a race condition during agent startup
where k8s node label updates are rejected
indefinitely by the host endpoint.

The current endpoint label update logic rejects
a request if any of the passed delete labels
are not present in the agent's endpoint state.

For host endpoints, the k8s node label add/update
events may be missed if the k8s node watcher
intializes before the host endpoint is created.
As a result, the host endpoint labels are outdated,
and all subsequent node label updates are rejected
due to the aforementioned precondition check.

To resolve the issue, this commit skips
checking the precondition for host endpoints.

Fixes #29649

Signed-off-by: Satish Matti <smatti@google.com>